Port remaining TransformStream tests to wpt
Move the remaining tests from test/transform-stream.js to
to-upstream-wpts/transform-streams and delete the old file.

Minimal porting has been done to make the tests work with
testharness.js. Conversion to using promise_test() has not yet been done
to make this change easier to review.

Changing the tests to use more idiomatic wpt style will be done in a
followup change.
